+++ Source/DotNetWorkQueue.IntegrationTests.Shared/SharedSetup.cs
@@ -28,14 +28,14 @@ namespace DotNetWorkQueue.IntegrationTests.Shared
             {
                 case InterceptorAdding.ConfigurationOnly:
                     return new QueueContainer<TTransportInit>(serviceRegister => serviceRegister.Register(() => metrics,
-                       LifeStyles.Singleton).Register(() => new AesMessageInterceptorConfiguration(System.Text.Encoding.ASCII.GetBytes("0123456789abcdef0123456789abcdef")                           ), LifeStyles.Singleton).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
+                       LifeStyles.Singleton).Register(() => new AesMessageInterceptorConfiguration(TestEncryptionKey.Aes256                           ), LifeStyles.Singleton).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
                 case InterceptorAdding.Yes:
                     return new QueueContainer<TTransportInit>(serviceRegister => serviceRegister.Register(() => metrics,
                         LifeStyles.Singleton).RegisterCollection<IMessageInterceptor>(new[]
                         {
                             typeof (GZipMessageInterceptor), //gzip compression
                             typeof (AesMessageInterceptor) //encryption
-                        }).Register(() => new AesMessageInterceptorConfiguration(System.Text.Encoding.ASCII.GetBytes("0123456789abcdef0123456789abcdef")                            ), LifeStyles.Singleton).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
+                        }).Register(() => new AesMessageInterceptorConfiguration(TestEncryptionKey.Aes256                            ), LifeStyles.Singleton).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
                 default:
                     return new QueueContainer<TTransportInit>(serviceRegister => serviceRegister.Register(() => metrics,
                         LifeStyles.Singleton).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
@@ -63,7 +63,7 @@ namespace DotNetWorkQueue.IntegrationTests.Shared
                                 .RegisterNonScopedSingleton(scope)
                                 .RegisterNonScopedSingleton(trace)
                                 .Register(() => metrics,
-                                    LifeStyles.Singleton).Register(() => new AesMessageInterceptorConfiguration(System.Text.Encoding.ASCII.GetBytes("0123456789abcdef0123456789abcdef")), LifeStyles.Singleton),
+                                    LifeStyles.Singleton).Register(() => new AesMessageInterceptorConfiguration(TestEncryptionKey.Aes256), LifeStyles.Singleton),
                             options => SetOptions(options, enableChaos));
                     case InterceptorAdding.Yes:
                         return new QueueContainer<TTransportInit>(
@@ -76,7 +76,7 @@ namespace DotNetWorkQueue.IntegrationTests.Shared
                                 {
                                     typeof(GZipMessageInterceptor), //gzip compression
                                     typeof(AesMessageInterceptor) //encryption
-                                }).Register(() => new AesMessageInterceptorConfiguration(System.Text.Encoding.ASCII.GetBytes("0123456789abcdef0123456789abcdef")), LifeStyles.Singleton),
+                                }).Register(() => new AesMessageInterceptorConfiguration(TestEncryptionKey.Aes256), LifeStyles.Singleton),
                             options => SetOptions(options, enableChaos));
                     default:
                         return new QueueContainer<TTransportInit>(
@@ -95,7 +95,7 @@ namespace DotNetWorkQueue.IntegrationTests.Shared
                     case InterceptorAdding.ConfigurationOnly:
                         return new QueueContainer<TTransportInit>(
                             serviceRegister => serviceRegister.Register(() => logProvider, LifeStyles.Singleton).Register(() => metrics,
-                                LifeStyles.Singleton).Register(() => new AesMessageInterceptorConfiguration(System.Text.Encoding.ASCII.GetBytes("0123456789abcdef0123456789abcdef")                                ), LifeStyles.Singleton).RegisterNonScopedSingleton(scope).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
+                                LifeStyles.Singleton).Register(() => new AesMessageInterceptorConfiguration(TestEncryptionKey.Aes256                                ), LifeStyles.Singleton).RegisterNonScopedSingleton(scope).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
                     case InterceptorAdding.Yes:
                         return new QueueContainer<TTransportInit>(
                             serviceRegister => serviceRegister.Register(() => logProvider, LifeStyles.Singleton).Register(() => metrics,
@@ -103,7 +103,7 @@ namespace DotNetWorkQueue.IntegrationTests.Shared
                             {
                                 typeof (GZipMessageInterceptor), //gzip compression
                                 typeof (AesMessageInterceptor) //encryption
-                            }).Register(() => new AesMessageInterceptorConfiguration(System.Text.Encoding.ASCII.GetBytes("0123456789abcdef0123456789abcdef")                                ), LifeStyles.Singleton).RegisterNonScopedSingleton(scope).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
+                            }).Register(() => new AesMessageInterceptorConfiguration(TestEncryptionKey.Aes256                                ), LifeStyles.Singleton).RegisterNonScopedSingleton(scope).RegisterNonScopedSingleton(trace), options => SetOptions(options, enableChaos));
                     default:
                         return new QueueContainer<TTransportInit>(
                             serviceRegister => serviceRegister.Register(() => logProvider, LifeStyles.Singleton).Register(() => metrics,